+++ apps/native-component-list/src/navigation/MainNavigators.tsx
@@ -3,6 +3,7 @@ import { DrawerNavigationOptions } from '@react-navigation/drawer';
 import { PathConfig } from '@react-navigation/native';
 import { StackNavigationOptions } from '@react-navigation/stack';
 import React from 'react';
+import { getScreenIdForLinking } from 'test-suite/screens/getScreenIdForLinking';
 
 import ExpoApisStackNavigator, { Screens as APIScreens } from './ExpoApisStackNavigator';
 import ExpoComponentsStackNavigator, {
@@ -30,7 +31,7 @@ const apis: ScreenConfig = {
       ...APIScreens.reduce(
         (prev, curr) => ({
           ...prev,
-          [curr.name]: curr.route || curr.name.toLowerCase(),
+          [curr.name]: getScreenIdForLinking(curr),
         }),
         {}
       ),
@@ -48,7 +49,7 @@ const components = {
       ...ComponentScreens.reduce(
         (prev, curr) => ({
           ...prev,
-          [curr.name]: curr.route || curr.name.toLowerCase(),
+          [curr.name]: getScreenIdForLinking(curr),
         }),
         {}
       ),