+++ packages/@expo/cli/src/start/server/metro/getCssModulesFromBundler.ts
@@ -1,6 +1,6 @@
 // NOTE(@kitten): jest-resolver -> resolve.exports bug (https://github.com/lukeed/resolve.exports/issues/40)
 import { getJsOutput, isJsModule } from '@expo/metro/metro/DeltaBundler/Serializers/helpers/js.js';
-import type { Module, ReadOnlyDependencies } from '@expo/metro/metro/DeltaBundler/types.flow';
+import type { Module, ReadOnlyDependencies } from '@expo/metro/metro/DeltaBundler/types';
 import type IncrementalBundler from '@expo/metro/metro/IncrementalBundler';
 import splitBundleOptions from '@expo/metro/metro/lib/splitBundleOptions';
 import type { MetroConfig } from '@expo/metro-config';