+++ packages/expo-dev-launcher/ios/EXDevLauncherController.m
@@ -40,9 +40,6 @@
 #define VERSION @ STRINGIZE2(EX_DEV_LAUNCHER_VERSION)
 #endif
 
-#define EX_DEV_LAUNCHER_PACKAGER_PATH @"packages/expo-dev-launcher/index.bundle?platform=ios&dev=true&minify=false"
-
-
 @interface EXDevLauncherController ()
 
 @property (nonatomic, weak) UIWindow *window;
@@ -91,106 +88,11 @@
   return self;
 }
 
-- (NSArray<id<RCTBridgeModule>> *)extraModulesForBridge:(RCTBridge *)bridge
-{
-  NSMutableArray<id<RCTBridgeModule>> *modules = [NSMutableArray new];
-
-  [modules addObject:[RCTDevMenu new]];
-#ifndef EX_DEV_LAUNCHER_URL
-  [modules addObject:[EXDevLauncherRCTDevSettings new]];
-#endif
-  [modules addObject:[DevClientNoOpLoadingView new]];
-
-  return modules;
-}
-
 + (NSString * _Nullable)version {
 #ifdef VERSION
   return VERSION;
 #endif
   return nil;
-}
-
-// Expo developers: Enable the below code by running
-//     export EX_DEV_LAUNCHER_URL=http://localhost:8090
-// in your shell before doing pod install. This will cause the controller to see if
-// the expo-launcher packager is running, and if so, use that instead of
-// the prebuilt bundle.
-// See the pod_target_xcconfig definition in expo-dev-launcher.podspec
-
-- (nullable NSURL *)devLauncherBaseURL
-{
-#ifdef EX_DEV_LAUNCHER_URL
-  return [NSURL URLWithString:@EX_DEV_LAUNCHER_URL];
-#endif
-  return nil;
-}
-- (nullable NSURL *)devLauncherURL
-{
-#ifdef EX_DEV_LAUNCHER_URL
-  return [NSURL URLWithString:EX_DEV_LAUNCHER_PACKAGER_PATH
-                relativeToURL:[self devLauncherBaseURL]];
-#endif
-  return nil;
-}
-
-- (nullable NSURL *)devLauncherStatusURL
-{
-#ifdef EX_DEV_LAUNCHER_URL
-  return [NSURL URLWithString:@"status"
-                relativeToURL:[self devLauncherBaseURL]];
-#endif
-  return nil;
-}
-
-- (BOOL)isLauncherPackagerRunning
-{
-  // Shamelessly copied from RN core (RCTBundleURLProvider)
-
-  // If we are not running in the main thread, run away
-  if (![NSThread isMainThread]) {
-    return NO;
-  }
-
-  NSURL *url = [self devLauncherStatusURL];
-  NSURLSession *session = [NSURLSession sharedSession];
-  NSURLRequest *request = [NSURLRequest requestWithURL:url
-                                           cachePolicy:NSURLRequestUseProtocolCachePolicy
-                                       timeoutInterval:1];
-  __block NSURLResponse *response;
-  __block NSData *data;
-
-  dispatch_semaphore_t semaphore = dispatch_semaphore_create(0);
-  [[session dataTaskWithRequest:request
-              completionHandler:^(NSData *d, NSURLResponse *res, __unused NSError *err) {
-                data = d;
-                response = res;
-                dispatch_semaphore_signal(semaphore);
-              }] resume];
-  dispatch_semaphore_wait(semaphore, DISPATCH_TIME_FOREVER);
-
-  NSString *status = [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding];
-  return [status isEqualToString:@"packager-status:running"];
-}
-
-- (NSURL *)sourceURLForBridge:(RCTBridge *)bridge
-{
-  return [self getSourceURL];
-}
-
-- (NSURL *)bundleURL
-{
-  return [self getSourceURL];
-}
-
-- (NSURL *)getSourceURL
-{
-  NSURL *launcherURL = [self devLauncherURL];
-  if (launcherURL != nil && [self isLauncherPackagerRunning]) {
-    return launcherURL;
-  }
-  NSURL *bundleURL = [[NSBundle mainBundle] URLForResource:@"EXDevLauncher" withExtension:@"bundle"];
-  return [[NSBundle bundleWithURL:bundleURL] URLForResource:@"main" withExtension:@"jsbundle"];
 }
 
 - (void)clearRecentlyOpenedApps
@@ -330,12 +232,6 @@
   [self.delegate destroyReactInstance];
 
 #if RCT_DEV
-  NSURL *url = [self devLauncherURL];
-  if (url != nil) {
-    // Connect to the websocket
-    [[RCTPackagerConnection sharedPackagerConnection] setSocketConnectionURL:url];
-  }
-
   [self _addInitModuleObserver];
 #endif