+++ packages/expo-server/src/ImmutableRequest.ts
@@ -52,7 +52,7 @@ export type _ImmutableRequest = Omit<
 /**
  * An immutable version of the Fetch API's [`Request`](https://developer.mozilla.org/en-US/docs/Web/API/Request) object which prevents mutations to the request body and headers.
  */
-export class ImmutableRequest implements _ImmutableRequest {
+export class ImmutableRequest implements _ImmutableRequest, RequestInit {
   readonly #headers: ImmutableHeaders;
   readonly #request: Request;
 
@@ -125,13 +125,12 @@ export class ImmutableRequest implements _ImmutableRequest {
     throw new Error('This operation is not allowed on immutable requests.');
   }
 
-  /**
-   * The request body is not accessible in immutable requests.
-   */
-  // @ts-expect-error This ensures JavaScript users cannot mutate the request body
-  // eslint-disable-next-line getter-return
-  get body() {
-    this.#throwImmutableBodyError();
+  /** The request body is not accessible in immutable requests. */
+  get body(): never {
+    // NOTE(@kitten): `new Request(req.url, req)` may internally access `req.body` to copy the request
+    // We can pretend it is `null`. Marking `bodyUsed` makes no sense here as it manipulates the subsequent
+    // code paths, but pretending there was no body should be safe
+    return null as never;
   }
 
   async arrayBuffer() {